修改登录时限和注册返回

+++ java/superblog-shiro/src/main/java/com/lhq/superboot/shiro/property/ShiroConfigProperty.java
@@ -47,7 +47,7 @@ public class ShiroConfigProperty {
 	private String sessionPrefix;
 
 	@Value("${shiro.session.valid}")
-	private int sessionValid;
+	private long sessionValid;
 	
 	@Value("${shiro.session.expire}")
 	private int sessionExpire;
